# Minor Project/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/api/admin/login": {
            "post": {
                "tags": [
                    "login admin"
                ],
                "summary": "Register new admin",
                "operationId": "login",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Updated name of the pet",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Updated name of the pet",
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/admin/register": {
            "post": {
                "tags": [
                    "Register admin"
                ],
                "summary": "Register new admin",
                "operationId": "register",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Updated name of the pet",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Updated name of the pet",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Updated name of the pet",
                                        "type": "password"
                                    },
                                    "password_confirmation": {
                                        "description": "Updated name of the pet",
                                        "type": "string"
                                    },
                                    "phone": {
                                        "description": "Updated name of the pet",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/admin/details": {
            "get": {
                "tags": [
                    "admin details"
                ],
                "summary": "admin details",
                "operationId": "details",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/logout": {
            "post": {
                "tags": [
                    "logout"
                ],
                "summary": "to logout admin",
                "operationId": "logout_admin",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/updater": {
            "post": {
                "tags": [
                    "update admin profile"
                ],
                "summary": "update admin profile details",
                "operationId": "update_profile",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Updated name of the pet",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Updated name of the pet",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Updated name of the pet",
                                        "type": "password"
                                    },
                                    "password_confirmation": {
                                        "description": "Updated name of the pet",
                                        "type": "string"
                                    },
                                    "phone": {
                                        "description": "Updated name of the pet",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/admin/users_list": {
            "get": {
                "tags": [
                    "user list"
                ],
                "summary": "get all the list of users",
                "operationId": "users_list",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/company_list": {
            "get": {
                "tags": [
                    "company list"
                ],
                "summary": "get all the list of companies",
                "operationId": "company_list",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/delete_user/{id}": {
            "post": {
                "tags": [
                    "delete user with id"
                ],
                "summary": "delete intern",
                "operationId": "delete_user_with_id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Status values that needed to be considered for filter",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "default": "available"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/delete_company/{id}": {
            "post": {
                "tags": [
                    "delete company with id"
                ],
                "summary": "delete a company by admin",
                "operationId": "delete_company_with_id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Status values that needed to be considered for filter",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "default": "available"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/delete_job_post/{job_id}": {
            "post": {
                "tags": [
                    "delete job_post with id"
                ],
                "summary": "delete a job_post by admin",
                "operationId": "delete_job_post_by_id",
                "parameters": [
                    {
                        "name": "job_id",
                        "in": "path",
                        "description": "Status values that needed to be considered for filter",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "default": "available"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/user/delete": {
            "post": {
                "tags": [
                    "delete intern"
                ],
                "summary": "delete intern",
                "operationId": "delete_user",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}